Refactor System.Win32.Async.IOManager and System.Wind32.Async.File
Allocation and deallocation of 'IOData' associated with each async
requests is done in Haskell.  We use `System.Win32` api to allocate
/ deallocate from windows process heap (as we did in C).

We can now use `ReadFile` and `WriteFile` syscalls without
C wrappers.  We are not using the ones from `Win32` package only beacuse
we support Win32<2.7 which did not defined OVERLAPPED data type.  This
can be changed later on.

'IOData' data type is holds 'OVERLAPPED' passed to system calls and the stable pointer.
As previously, this allows 'HsGetQueuedCompletionStatus' to use
'CONTAINING_RECORD' macro to recover it just from 'OVERLAPPED' member.

'withIODataPtr' is introduced which allocats 'IOData' on process heap.
It takes most of responsibilities of 'waitForCompletion' function, which
has been removed.  This will allow us to add a 'CacnelIoEx' call in case
the async I/O was interrupted by a GHC'c async exception.