withIOManager use IOException
Before this patch `IOCompletionException` thrown by io-manger thread
weren't re-throw to the applicaiton thread by `withIOManager`; This is
fixed in this patch.  Instead of dedicated `IOCompetionException` for
the `dequeueCompletionPackets` we just use `IOException`.

Also improve the documentation why we are not throwing when
'c_GetQueuedCompletion' errors with an 'ERROR_INVALID_HANDLE'.